Update README.md
I have been given this part of the empire to manage, and I see it is... a complicated inheritance. It is not what I would have built myself, but it is my duty to work with it, not against it. To complain of its nature is to forget that my own virtue lies in my handling of what I am given.

Signed-off-by: James Ross <james@flyingrobots.dev>